backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv

python-multipart
passlib[bcrypt]
bcrypt

pytz
python-dateutil

pillow
pytesseract
pdfplumber
pdf2image
PyPDF2
python-docx

reportlab

apscheduler

requests

transformers
torch
sentencepiece
yake

flask
flask-cors

email-validator

apscheduler
pytz
